🐛🤖 Cap per-phase reminders at min(7 days, cycle duration) — slashbin review #4
`reminderValue` on pricing phases had no upper bound: a reminder longer
than the phase's own cycle fired *before* the phase even started, and
anything over 7 days silently overshot the cron's scan window and never
went out. Aligning the phase reminder with the same 7-day ceiling as
the product-level reminder keeps the cron behaviour consistent (we'll
revisit widening the window in a separate ticket) and adding the
`reminder ≤ cycle` invariant prevents the pathological "renew now"
race.

Enforced both server-side (Zod `.refine()` on each phase) and in the
admin form (reactive `max` on the reminder input, recomputed from the
phase's unit and reminder unit).

Also move `subscriptionUnitToSeconds` from `$lib/server/subscriptions`
to `$lib/types/SubscriptionDuration` so the admin form can call the
same mapping the schema uses — no more parallel implementations to
drift.

Refs #2651
